Add support for `ON CONFLICT DO UPDATE`
With this commit, we support the majority of PG's upsert. The things
that are still missing are things I want to support at some point, but
they're lower priority. The things that are known to be missing are:

- Support for expression indexes
    - We need a way for the entire ast to know not to qualify column
      names in that one position
- Where clauses on the `DO UPDATE` clause
    - Should actually be easy to do, just not high priority
- Support for using `excluded.column` in arbitrary expressions
    - This probably already works, just needs tests. Additional work is
      needed to make things like the `+` operator work though.

`ON CONFLICT (target) DO UPDATE SET changes` is done as:
`insert(&record.on_conflict(target, do_update().set(changes)))`. Changes
can be anything that works with normal update statements. You can
reference the value that was proposed for insertion as
`excluded(column)` (this is important for multi-row insertions).

There's one notable type safety hole which is that `excluded` can be
passed to normal update statements and compile successfully. We should
fix this at some point, but it's not worth blocking this feature for,
as you'd only ever really run into it by explicitly trying to break
things.

Fixes #196.