Document some of the generated items in the `table!` macro
I've had to be fairly generic in the generated docs, since we can't
actually pull in any of the macro variables. I've purposely left the
module itself and the column structs undocumented as those would come
from the doc comments given by the caller of the macro. I've left the
handling of doc comments for separate PR, to avoid conflicting with #779.

There's not really any way that we can test this, so the tests are
unaffected.