Make `AstPass` an opaque struct
I'm moving towards merging `to_sql` with `walk_ast`. The code will look
*really* bad if there's a ton of conditionals on which pass it is.
Instead I'd like to have the code look the same as it would if I were
able to make `AstPass` a trait (see 91c9d5a8ebdb67fdcd116e024793ee3 for
context on why we can't do that)

In order to enforce that only the methods are used, and nothing is
actually destructuring, I've had to do a bit of funky indirection.