Allow bind parameters to be supplied to raw SQL queries
This is a first pass at a much requested feature. While this does not
provide the richer raw SQL story I eventually want to provide, this
should serve as a sufficient stopgap. Bind parameters are included in
the raw SQL string, and then bound after the fact by calling `.bind`
once for each parameter (supplying the SQL type of the parameter at the
call site).

It should be noted that I have explicitly *not* implemented `Expression`
for the new AST node that this added, as it seems like it'll be too easy
to mess up the parameter index when it's used as a fragment of a query
constructed using the query builder. This can be expanded in the future
(maybe only allowed on SQLite and MySQL which don't index bind params),
but I wanted to be conservative to start.