Pass around `&DB::TypeMetadata` to where it's eventually needed
The main goal of this change is to have a reference to
`DB::TypeMetadata` in the metadata function. This change wires up
everything needed to actually have that reference, and pass it around to
where it's needed. For now the concrete type is always still `()`, but
that will change in the next commit.