Don't implement expression methods on tuples
With the exception of `.nullable`, all of these methods would generate
invalid SQL if used. Some people mistakenly think that we map rust
tuples to a PG tuple. We don't actually support PG tuples, and if we did
we'd need to have it be a separate type to differentiate `SELECT id,
name` from `SELECT (id, name)`.

This is not a full solution to the problem. Ideally we would actually
properly represent in the type system that a tuple is a list, which is
only valid in a few contexts. Instead, this solution simply prevents
their construction. I don't think this mistake is common enough to
warrant more time being spent on it.

My hope is that if/when variadic generics happen, they are not built
around tuples, which would allow us to better separate this case from
normal values. Until that time, this should suffice.

I'm not 100% sold on the name `SingleValue`, as it seems odd that
`Array<T>` implements it. However, the only other name I could think of
was `NotTuple`, which will feel odd if we ever do support PG tuples.
Something like `NotColumnList` could work, but that name feels odd to
me. I think I want to stick with a name that says what the type is, not
what it isn't.

Fixes #998.