Fix intermittent deadlock in PG tests
The issue occurred because the following events happen:

- Transaction 1 begins. It acquires an exclusive lock on its transaction
  id
- Transaction 2 begins. It acquires an exclusive lock on its transaction
  id
- Transaction 1 does `INSERT INTO users (id, name) VALUES (1, 'Sean'),
  (2, 'Tess')`. It acquires a row lock on the inserted rows. The insert
  succeeds.
- Transaction 2 does the same insert. It acquires a row lock on the
  inserted rows. The success of the insert depends on whether
  transaction 1 commits or not, so it attempts to acquire a share lock
  on transaction 1's id. This conflicts with the exclusive lock in step
  1. The lock is not granted, and the transaction is blocked.
- Transaction 1 does `CREATE UNIQUE INDEX ON users (name) WHERE name !=
  'Tess'`. This requires a share lock on the table, which conflicts with
  the row lock in the previous step. The lock is not granted, and the
  transaction is blocked.
- Deadlock timeout is reached

Honestly, it seems like transaction 2 shouldn't be getting the row lock
until it's able to continue. I can't really fix PG though, so for the
time being we need to work around it. The simplest solution (and the one
in this commit) is to just acquire the table lock before the row lock.
Long term, we should really fix our tests to not rely on the specific
values of the IDs. This would cause the deadlock to not occur
(transaction 2 wouldn't have been blocked), and would let more of our
tests run in parallel.