Bump recursion limit in diesel_tests
The attribute in `schema.rs` didn't seem to do anything. I'm not sure if
that is a regression in Rust or an oversight on our part (and the attr
wasn't necessary before).

Additionally, looking at <https://github.com/rust-lang/rust/pull/41676>,
this shouldn't be necessary at all in recent nightlies (after May 7),
but without the explicit setting, it fails to compile and suggests
adding `#![recursion_limit="128"]`. If I read the code in [1] correctly,
this means it was set to 64 before.

[1]: https://github.com/rust-lang/rust/blob/5c7add7551fa093f655e76650a48564ac347c83a/src/librustc/traits/error_reporting.rs#L1150-L1156